Backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-crontab==0.7.1
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
jsonfield==3.1.0
packaging==24.2
pillow==10.4.0
psycopg2==2.9.9
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
